terminada bazooka + sus constantes

+++ src/codigo/TP taller/TP taller/Modelo/Armas/Aleluya.h
@@ -9,7 +9,7 @@ public:
 	Aleluya(void);
 	Aleluya(float x, float y, short int rotacion, b2World* world, bool estatico, float radioExplosion, float radioArma, float masa);
 	~Aleluya(void);
-	void disparar(bool sentido, int potencia, float angulo);
+	void disparar(bool sentido, float potencia, float angulo);
 };
 
 #endif
